+++ codegen/codegen.go
@@ -28,7 +28,7 @@ import (
 	"github.com/getkin/kin-openapi/openapi3"
 	"golang.org/x/tools/imports"
 
-	"github.com/discord-gophers/goapi-gen/pkg/codegen/templates"
+	"github.com/discord-gophers/goapi-gen/templates"
 )
 
 // Options defines the optional code to generate.
