+++ codegen/inline.go
