+++ codegen/operations.go
