+++ codegen/prune.go
