+++ codegen/prune_test.go
