+++ examples/petstore-expanded/petstore.go
@@ -14,7 +14,7 @@ import (
 	"github.com/go-chi/chi/v5"
 
 	"github.com/discord-gophers/goapi-gen/examples/petstore-expanded/api"
-	"github.com/discord-gophers/goapi-gen/pkg/middleware"
+	"github.com/discord-gophers/goapi-gen/middleware"
 )
 
 func main() {
