+++ internal/test/parameters/parameters.gen.go
@@ -15,7 +15,7 @@ import (
 	"path"
 	"strings"
 
-	"github.com/discord-gophers/goapi-gen/pkg/runtime"
+	"github.com/discord-gophers/goapi-gen/runtime"
 	"github.com/getkin/kin-openapi/openapi3"
 	"github.com/go-chi/chi/v5"
 	"github.com/go-chi/render"
