+++ internal/test/parameters/parameters_test.go
@@ -9,7 +9,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/discord-gophers/goapi-gen/pkg/testutil"
+	"github.com/discord-gophers/goapi-gen/internal/testutil"
 )
 
 type testServer struct {
