+++ templates/enum-typedef.tmpl
