+++ templates/imports.tmpl
@@ -20,8 +20,8 @@ import (
 	"strings"
 	"time"
 
-	"github.com/discord-gophers/goapi-gen/pkg/runtime"
-	openapi_types "github.com/discord-gophers/goapi-gen/pkg/types"
+	"github.com/discord-gophers/goapi-gen/runtime"
+	openapi_types "github.com/discord-gophers/goapi-gen/types"
 	"github.com/getkin/kin-openapi/openapi3"
 	"github.com/go-chi/chi/v5"
 	"github.com/go-chi/render"
