+++ templates/typedef.tmpl
